+++ packages/card/src/Card/styles.ts
@@ -3,8 +3,7 @@ import { Flex } from '@siakit/layout'
 
 export const CardContainer = styled(Flex, {
   backgroundColor: '$shape',
-  boxShadow: '$sm',
 
   borderRadius: '$md',
-  border: '1px solid $gray3',
+  border: '1px solid $gray4',
 })