+++ packages/avatar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @siakit/avatar
 
+## 0.0.2
+
+### Patch Changes
+
+- various fix
+- Updated dependencies
+  - @siakit/heading@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes