+++ packages/page-header/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @siakit/page-header
 
+## 0.0.2
+
+### Patch Changes
+
+- various fix
+- Updated dependencies
+  - @siakit/card@0.0.3
+  - @siakit/heading@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes